services/order/pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.0"
description = ""
authors = ["Hugo Brilhante <hugobrilhante@gmail.com>"]
readme = "README.md"
packages = [{include = "template", from = "src"}]

[tool.poetry.dependencies]
python = "3.12.1"
django = "^5.0.1"
django-configurations = {extras = ["cache", "database", "email", "search"], version = "^2.5"}
whitenoise = "^6.6.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
python-json-logger = "^2.0.7"
djangorestframework = "^3.14.0"
django-outbox-pattern = "^1.0.1"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "3.5.0"

[tool.ruff]
line-length = 120
include = ["pyproject.toml", "src/**/*.py", "test/**/*.py"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]


[tool.ruff.lint.isort]
force-single-line = true
force-sort-within-sections = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
